Reapply accidently reverted printf format fix from shadeslayer

+++ pcp_app/pcp_app.c
@@ -359,11 +359,11 @@ static void print_get_dscp(pcp_flow_t* f)
      for (; cnt>0; cnt--, ret++) {
          char ntop_buff[INET6_ADDRSTRLEN];
          char timebuf[32];
-         printf("%-20s %5hu %3d %5s %s",
+         printf("%-20s %5d %3d %5s %s",
                  inet_ntop(AF_INET6, &ret->int_ip, ntop_buff,
                      sizeof(ntop_buff)),
-                 ret->learned_dscp,
-                 ret->pcp_result_code,
+                 (int)ret->learned_dscp,
+                 (int)ret->pcp_result_code,
                  decode_fresult(ret->result),
                  ret->recv_lifetime_end == 0 ? " -\n" :
                          ctime_r(&ret->recv_lifetime_end, timebuf)); //LCOV_EXCL_LINE
